requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.0
numpy==1.24.3
gunicorn==21.2.0 